Return UINT instead of size_t

+++ src/Cpp/1-getting-started/1-2-3-NamingThings/ShaderCollection.cpp
@@ -46,12 +46,12 @@ ShaderCollection ShaderCollection::CreateShaderCollection(const ShaderCollection
     return collection;
 }
 
-size_t ShaderCollection::GetLayoutByteSize(VertexType vertexType)
+UINT ShaderCollection::GetLayoutByteSize(VertexType vertexType)
 {
     switch (vertexType)
     {
     case VertexType::PositionColor:
-        return sizeof(VertexPositionColor);
+        return static_cast<UINT>(sizeof(VertexPositionColor));
     }
     return 0;
 }
